+++ NEWS.md
@@ -102,3 +102,4 @@
 # odytools dev
 
 -   Now, project-specific functions of a RedCap project are assumed to be in the scripts of the folder functions.
+-   Added **ody_rc_timatravel** to load previous imports.